# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Remote Dev API
  description: |
    API for Remote Dev - a web-based terminal interface with persistent sessions,
    GitHub integration, and session management features.
  version: 1.0.0
  contact:
    name: Remote Dev
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: http://localhost:6001
    description: Production server
...
paths:
...
  /api/groups:
    get:
      tags: [Groups]
      summary: List groups
      responses:
        '200':
          description: List of project groups
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'

    post:
      tags: [Groups]
      summary: Create group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - parentGroupId
              properties:
                name:
                  type: string
                parentGroupId:
                  type: string
                  format: uuid
                  nullable: true
                sortOrder:
                  type: integer
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                type: object
                properties:
                  group:
                    $ref: '#/components/schemas/Group'
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List projects
      parameters:
        - name: groupId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter to one group; omit to list every project the user owns
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'

    post:
      tags: [Projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - groupId
                - name
              properties:
                groupId:
                  type: string
                  format: uuid
                name:
                  type: string
                sortOrder:
                  type: integer
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: '#/components/schemas/Project'
...
components:
  schemas:
...
    Group:
      type: object
      description: Project group (container; preferences only)
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        parentGroupId:
          type: string
          format: uuid
          nullable: true
        collapsed:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Project:
      type: object
      description: Project (leaf; owns sessions, tasks, channels, secrets)
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        name:
          type: string
        collapsed:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
security:
  - sessionCookie: []